programador/requirements.txt
flask
flask-wtf
google-cloud-firestore
google-cloud-tasks
esptool
gunicorn
google-cloud-storage
google-cloud-run
flask_socketio